solution/services/book-service/requirements.txt
fastapi
click
uvicorn[standard]
sqlalchemy
asyncpg>=0.27.0
python-dotenv
python-jose[cryptography]
cachetools
httpx
